pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["youngdon <ydkim@stealien.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.28.0"
debugpy = "^1.8.11"
minio = "^7.2.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
